refactor: JdbcConnection 클래스 분리
