+++ src/main/java/chess/dao/BoardDao.java
@@ -1,6 +1,5 @@
 package chess.dao;
 
-import java.sql.Connection;
 import java.util.Map;
 
 public interface BoardDao {
@@ -9,6 +8,4 @@ public interface BoardDao {
     Map<String, String> findById(int gameId);
 
     void deleteById(int gameId);
-
-    Connection getConnection();
 }
