+++ src/main/java/chess/dao/BoardDaoImpl.java
@@ -1,7 +1,8 @@
 package chess.dao;
 
+import static chess.dao.JdbcConnection.*;
+
 import java.sql.Connection;
-import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -10,14 +11,10 @@ import java.util.Map;
 
 public class BoardDaoImpl implements BoardDao {
 
-    private static final String URL = "jdbc:mysql://localhost:13306/chess";
-    private static final String USER = "user";
-    private static final String PASSWORD = "password";
-
     @Override
     public void save(int gameId, Map<String, String> boardMap) {
-        Connection connection = getConnection();
         String sql = "insert into board (game_id, position, piece) values (?, ?, ?)";
+        Connection connection = getConnection();
         try {
             PreparedStatement statement = connection.prepareStatement(sql);
             for (Map.Entry<String, String> entry : boardMap.entrySet()) {
@@ -63,16 +60,5 @@ public class BoardDaoImpl implements BoardDao {
         } catch (SQLException e) {
             e.printStackTrace();
         }
-    }
-
-    @Override
-    public Connection getConnection() {
-        Connection connection = null;
-        try {
-            connection = DriverManager.getConnection(URL, USER, PASSWORD);
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        return connection;
     }
 }
