+++ src/main/java/chess/dao/GameDao.java
@@ -1,12 +1,10 @@
 package chess.dao;
 
-import java.sql.Connection;
 import java.util.List;
 
 import chess.Game;
 
 public interface GameDao {
-    Connection getConnection();
 
     void save(Game game);
